Added draft probes.ts file in e2e directory as per issue
https://github.com/responsible-ai-collaborative/aiid/issues/2886

The probe tests in probes.ts use the (new) pages-tests.ts/testUrls
utility function (which is a generalized version of the previous
pages-tests.ts/testPages.

I've tried to have these tests be somewhat similar to the tests in
pages?.spec.ts.  Someone may want to refactor the tests for organization
and desired consistency, I'm +1 for any renaming/refactoring/etc that is
desired.

I don't know how/when it's desired to run these tests, and so need to
consult with team about that before changing the config and/or github
actions appropriately.  This is